refactor(tableGroup): 별도 패키지 분리
