+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,9 +1,13 @@
-package kitchenpos.table.application;
+package kitchenpos.tablegroup.application;
 
+import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.*;
-import kitchenpos.table.dto.TableGroupRequest;
-import kitchenpos.table.dto.TableGroupResponse;
-import kitchenpos.table.exception.NotFoundTableGroupException;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
+import kitchenpos.tablegroup.exception.NotFoundTableGroupException;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupCreatingValidator;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,22 +16,22 @@ import java.util.List;
 @Service
 @Transactional(readOnly = true)
 public class TableGroupService {
-    private final TableDomainService tableDomainService;
+    private final TableService tableService;
     private final TableGroupCreatingValidator tableGroupCreatingValidator;
     private final TableGroupRepository tableGroupRepository;
 
-    public TableGroupService(TableDomainService tableDomainService,
+    public TableGroupService(TableService tableService,
                              TableGroupCreatingValidator tableGroupCreatingValidator,
                              TableGroupRepository tableGroupRepository) {
-        this.tableDomainService = tableDomainService;
+        this.tableService = tableService;
         this.tableGroupCreatingValidator = tableGroupCreatingValidator;
         this.tableGroupRepository = tableGroupRepository;
     }
 
     @Transactional
     public TableGroupResponse create(TableGroupRequest request) {
-        List<OrderTable> orderTables = tableDomainService.findAllIdIn(request.toOrderTableIds());
-        tableGroupCreatingValidator.validate(request, orderTables);
+        List<OrderTable> orderTables = tableService.findAllIdIn(request.toOrderTableIds());
+        tableGroupCreatingValidator.validate(request.toOrderTableIds(), orderTables);
         TableGroup tableGroup = tableGroupRepository.save(new TableGroup());
         tableGroup.addOrderTables(orderTables);
         return TableGroupResponse.of(tableGroup);
