+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupUngroupedEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.util.Collections;
 import java.util.List;
