+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -1,6 +1,7 @@
-package kitchenpos.table.dto;
+package kitchenpos.tablegroup.dto;
 
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
