+++ src/main/java/kitchenpos/tablegroup/exception/NotFoundTableGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.exception;
+package kitchenpos.tablegroup.exception;
 
 public class NotFoundTableGroupException extends RuntimeException {
