+++ src/test/java/kitchenpos/table/domain/FixtureTableGroup.java
@@ -1,5 +1,7 @@
 package kitchenpos.table.domain;
 
+import kitchenpos.tablegroup.domain.TableGroup;
+
 public class FixtureTableGroup extends TableGroup {
 
     public FixtureTableGroup() {
