+++ src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.table.domain;
 
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
